requirements.txt
mysql-connector==2.2.9
et-xmlfile==1.1.0
openpyxl==3.0.10
aiohttp==3.8.1
pandas==1.3.5
requests~=2.27.1
python-dotenv==0.20.0